dash sandbox

+++ src/book/types.h
@@ -12,9 +12,9 @@ namespace liquibook { namespace book {
   typedef uint64_t Price;
   typedef uint64_t Quantity;
   typedef uint64_t Cost;
-  typedef uint32_t FillId;
-  typedef uint32_t ChangeId;
-  typedef uint32_t OrderConditions;
+  typedef uint64_t FillId;
+  typedef uint64_t ChangeId;
+  typedef uint64_t OrderConditions;
 
   enum OrderCondition {
     oc_no_conditions = 0,
